Continue inspection when some resources are not found
Purpose of inspection is to get whatever data it can in best effort.
In case of some resources are not found, inspect still can continue
for other resources.

This PR handles isnotfound error for resources and continues
to inspect. After inspection is completed, it prints error message
to warn user about non-existed resources.