must-gather debug pods should have priority class set
`oc` uses "system-cluster-critical" for the must-gather pods it
starts[1], but when the scripts in must-gather launch debug pods for log
collection[2][3], these are missing a priority class.

This PR ensures that debug pods in the openshift-must-gather namespace
have a priority class set. It's likely that these debug pods from
must-gather are collecting data from nodes that may be irrecoverable at
a later time and when using must-gather in this way, it's fairly likely
that these debug pods will be running on nodes that are under stress.

We also have a test that actually tries to make sure all pods in
openshift-* namespaces have a priority class, and if it happens to run
at the same time we're testing must-gather, we get test failures[4].

[1] https://github.com/openshift/oc/blob/master/pkg/cli/admin/mustgather/mustgather.go#L686-L689
[2] https://github.com/openshift/must-gather/blob/master/collection-scripts/gather_core_dumps#L11
[3] https://github.com/openshift/must-gather/blob/master/collection-scripts/gather_multus_logs#L13
[4] https://search.ci.openshift.org/?search=pods+found+with+invalid+priority+class&maxAge=168h&context=1&type=bug%2Bjunit&name=&excludeName=&maxMatches=5&maxBytes=20971520&groupBy=job