pkg/cli/admin/upgrade: Clarify client-side vs. server-side docs
The outgoing docs were frequently conflated.  The incoming docs
hopefully make the distinction clearer, and they list the client-side
checkForUpgrade conditions.  I haven't mentioned the Invalid check,
because that trips so rarely, but I have used the open-ended "include
checks for...", so folks aren't too surprised if they hit a
client-side guard that's not listed in the docs.  The "failing
clusters" bit is a lie, but that's already tracked in [1], and
eventually I'll be able to talk someone into reviewing the fix.

The cluster-side guards are also open-ended, although I do call out
both release verification (signature checks and similar) and
upgradeable conditions, since those are declared in the force godocs
[2].  I do not mention additional guards like the etcd backups [3],
because the version of oc requesting the update may diverge from the
version of the cluster being asked to update, so we don't really know
which checks that cluster will run cluster-side.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1992680
[2]: https://github.com/openshift/api/blob/22eb4f6f4385a0183a5eee4c8ca6d49eecda8120/config/v1/types_cluster_version.go#L378-L386
[3]: https://bugzilla.redhat.com/show_bug.cgi?id=1997347