pkg/cli/admin/release/extract: Add --included and --install-config
There are many manifests in release images.  CredentialsRequests,
Deployments, ClusterOperators, etc., etc.  There are increasingly
complicated conditions on whether the cluster-version operator
reconciles or ignores a manifest, including cluster profiles, cluster
capabilities, and feature sets [1].  To make it easier for users to
see the manifests that actually matter for their cluster, this commit
adds --included, which uses the same Manifest.Include logic used by
the cluster-version operator [2].  As long as we keep the vendored
copies of library-go close-ish between oc and the CVO, this will allow
users to just get the manifests that apply to their cluster.  For
example:

a. You have enabled TechPreviewNoUpgrade, so here's a
   CredentialsRequest for the openshift-cluster-api namespace [3], or
b. You have not enabled TechPreviewNoUpgrade, so I'll filter out
   this CredentialsRequest for the openshift-cluster-api namespace.

This should work fairly well as long as 'oc' can reach the cluster to
retrieve the same configuration data that the cluster-version operator
is using to make its decisions, although there is a risk of drift
between the separate code bases.  It gets trickier when we need to run
before the cluster exists.  For those cases, we have --install-config,
and 'oc' can get some information reliably from there (enabled
capability (sets), cloud provider, etc.) but not others (known
capabilities, cluster-profile, etc.)  We may need to circle back later
with additional flags for these properties, so the user can explicitly
pass in the information that we cannot figure out from the
install-config alone.

[1]: https://github.com/openshift/enhancements/blob/fb6021978e149e4cd12a223134a148172eff51ad/dev-guide/cluster-version-operator/dev/operators.md#manifest-inclusion-annotations
[2]: https://github.com/openshift/cluster-version-operator/blob/4eb9433f345cc551c7d31996ee441d649f677b4e/pkg/cvo/sync_worker.go#L963
[3]: https://github.com/openshift/cluster-capi-operator/blob/1593e40d6422e46473beee2f4419514a0418da1c/manifests/0000_30_cluster-api_00_credentials-request.yaml#L9-L15