pkg/cli/login: Warn, but do not fail, on surprise project-list errors
Initially, all project-listing errors were fatal.  bcf5470fe7 (make oc
login tolerate kube with --token, 2016-09-27) began tolerating
IsNotFound.  40e76d093d (make login, project, and discovery work
against kube with RBAC enabled, 2016-10-12) began tolerating
IsForbidden.  But sometimes the error can be on the service side, like
[1]:

  $ oc login --token=...

  Logged into "https://api..." as "..." using the token provided.

  Error from server (Timeout): the server was unable to return a response in the time allotted, but may still be processing the request (get projects.project.openshift.io)

With this commit, we log the error, so the caller is aware that we had
problems, but we fall back to 'default', because it's hard to imagine
a situation where the caller would rather not be logged in.  If they
want to adjust their default project after logging in, they can always
do that in follow-up work.

[1]: https://issues.redhat.com/browse/OCPBUGS-2495