pkg/cli/admin/upgrade: Drop "force" from "No updates available"
The outgoing wording dates back to the initial subcommand
implementation in openshift/origin@65cce8c37d (Add `oc adm upgrade` to
display available updates or trigger an update, 2018-12-04,
openshift/origin#21605).  But "force" is easy to conflate with the
--force option, and we occasionally hear of users who set --force
under the mistaken impression that it is related to the
recommended-ness of the target release image.

This message is only rendered when availableUpdates is empty, so
--to-image is unlikely to be sufficient.  The current output will be:

  $ oc adm upgrade --to-image quay.io/openshift-release-dev/ocp-release@sha256:20e270c3349fe2fcb38fd0da155329babc02d6b53e7e06ff235346c3c1cf11b5
  error: no recommended updates, specify --allow-explicit-upgrade to continue with the update or wait for new updates to be available.

which we could stiffen up by including some of --help's current
warnings about --allow-explicit-upgrade risk, but we can address that
in follow-up work.