Remove must-gather pod after it completes when explicit namespace is used
Currently, when user passes `--run-namespace` flag, must-gather pod
becomes stale and is not removed.

The reason is that `copy` container in the pod is set sleeping
infinite, thus pod never falls into `completed` status to be collected
by garbage collector.

One option is setting to a precise duration for sleeping. But that
also means that we will delay pod removal to a certain time which is
presumably long(2 hours maybe?).

Other option is removing the must-gather pod manually.

This PR applies second option because we know when task is completed
and there might be no need to delay this removal.