pkg/cli/admin/release/mirror: Reject by-digest --to-release-image pullspecs
'oc adm release mirror ...' should tag both the main release image and
all the referenced component images when it pushes to the registry.
Ideally all downstream consumers are using by-digest pullspecs,
because those are immutable, and therefore safer, as described in
43cb3c8598 (pkg/cli/admin/upgrade: Client-side by-tag guard,
2020-04-16, #390).  But many image registries garbage-collect untagged
images, so when we push into a registry we need to push into a tag to
avoid that garbage collection.