[OCPNODE-1580]update oc amd release mirror output using idms
Add --print-mirror-instructions flag to command oc adm release.
The print out instructions of the command will keep it as is if the flag is unset or set to `icsp`.
If set to 'idms', update instruction output of `oc adm release mirror` to ImageDigestMirrorSet example.
The new ImageDigestMirrorSet examples are:
- use `imageDigestSources` field in install-config.yaml[1] that will creates ImageDigestMirrorSet manifests,
- manifest example of ImageDigestMirrorSet CRD

[1] Add `imageDigestSources` to install-config: https://github.com/openshift/installer/pull/6235

result of this patch:
```
$ oc adm release mirror --print-mirror-instructions=idms  -a /home/qiwan/cluster/mysecret --from=quay.io/openshift-release-dev/ocp-release:4.13.2-x86_64 --to=localhost:5000/ocp4/openshift4 --to-release-image=localhost:5000/ocp4/openshift4:4.13-x86_64 --dry-run

To use the new mirrored repository to install, add the following section to the install-config.yaml:

imageDigestSources:
- mirrors:
  - localhost:5000/ocp4/openshift4
  source: quay.io/openshift-release-dev/ocp-release
- mirrors:
  - localhost:5000/ocp4/openshift4
  source: quay.io/openshift-release-dev/ocp-v4.0-art-dev

To use the new mirrored repository for upgrades, use the following to create an ImageDigestMirrorSet:

apiVersion: config.openshift.io/v1
kind: ImageDigestMirrorSet
metadata:
  name: example
spec:
  imageDigestMirrors:
  - mirrors:
    - localhost:5000/ocp4/openshift4
    source: quay.io/openshift-release-dev/ocp-release
  - mirrors:
    - localhost:5000/ocp4/openshift4
    source: quay.io/openshift-release-dev/ocp-v4.0-art-dev

```

Signed-off-by: Qi Wang <qiwan@redhat.com>