Always copy the blobs if the target isn't a registry
The assuption that the manifest existing means that the blobs exist only works when the target is a registry that will validate that the blobs have been uploaded before the manifest. On other targets, we should check for layer existence each time because the user may have used "--continue-on-error" and end up with missing layers.