ocpcertificates: complete MCS CA rotation command
Add a new subcommand to update cert references. This can be run
separately but is also (defaults to true) a flag in the cert rotation
path.

The two commands are:

1. regenerate-machine-config-server-serving-cert
This creates a new MCS CA and cert

2. update-ignition-ca-bundle-for-machine-config-server
This updates the ignition in user-data (machine-api namespace) to
reference the new cert, so new nodes are still able to join the cluster.

Note that for some UPI/non-machineset-backed setups, users will have to
manually rotate the cert in an external location.