pkg/cli/admin/update: Rename "upgrade" to "update" and similar
The outgoing "upgrade" phrasing still works, but now logs deprecation
warnings.  We may never remove the outgoing "upgrade" phrasing,
because it seems unlikely to be worth the compatibility churn.  But
moving the canonical phrasing to "update" aligns with
openshift/openshift-docs@a68b607325 (Adding 'update' to the glossary,
2022-01-22, openshift/openshift-docs#40525).  And logging the
deprecation warnings makes the new canonical phrasing more obvious,
which should help gradually transition existing consumers, because the
more consistently folks use the canonical phrasing, the easier it is
to find it with grep-searches when searching scripts and similar.