Add Windows node functionality to debug command
Allows oc debug to be ran against a Windows node. Adds Windows specific
fields to the pod spec, if the node is detected to be Windows.

When debugging a Windows node the image option is required as we cannot
ship Windows images as part of the OpenShift payload. This means we
cannot set a default image, and the user must provide their own.