Revert "pkg/cli/admin/release/extract: Add --included and --install-config"
This reverts commit 9ec4f46e2fa3110f558900d5c482fb3b448bbc5d.

We suspect that https://github.com/openshift/oc/pull/1521 has broken all
Metal jobs, an example of a failure:
https://prow.ci.openshift.org/view/gs/origin-ci-test/pr-logs/pull/openshift_cluster-baremetal-operator/355/pull-ci-openshift-cluster-baremetal-operator-master-e2e-metal-ipi-ovn-ipv6/1691359315740332032.

The testing scripts we use set KUBECONFIG in advance to the location
where we'll create it. At the time "oc adm extract" is called, the file
does not exist yet. While you could argue that we should not do it, it
has worked for years, and it's quite possible that customers have
similar automation (e.g. setting KUBECONFIG as a global variable in
their playbooks). In any case, I don't think "oc adm extract" should try
to read the configuration if it does not explicitly need it.