pkg/cli/admin/release/extract_tools: Enable ImageRegistry, etc. on 4.13-to-4.14
Catching up with 4.14's [1].  As with MachineAPI, all of the new
capabilities were enabled in all 4.13 clusters, and because
capabilities cannot be uninstalled [2], that means they'll be enabled,
possibly implicitly, after an update to 4.14.  Including them here
ensures that we extract the ImageRegistry CredentialsRequest in:

  $ oc adm release extract --included --credentials-requests --to credentials-requests quay.io/openshift-release-dev/ocp-release:4.14.0...

when connecting to a currently-4.13 cluster.  [3] is tracking a more
sustainable long-term approach.

[1]: https://github.com/openshift/cluster-version-operator/pull/950/commits/ba3aeb9ab9d6d20a7355f80db4c47950b307a893#diff-f3e6bd2d0c764f1c28fd5d20520a2224b1f3cb5306500aeefd056a4efc857d34R318-R336
[2]: https://github.com/openshift/enhancements/blob/6cebc13f0672c601ebfae669ea4fc8ca632721b5/enhancements/installer/component-selection.md#capabilities-cannot-be-uninstalled
[3]: https://issues.redhat.com/browse/OTA-1010