OCPBUGS-20474: Set ImportPolicy to PreserveOriginal to honor --keep-manifest-list when mirroring a payload to an image stream
When mirroring a multiarch payload via `oc adm release mirror --from=... --to-image-stream=release --keep-manifest-list=true`, the ImageStreamTags are created with the default ImportPolicy.ImportMode (Legacy). Therefore, the release is not mirrored as multiarch/manifest-list, but will be mirrored as a single-arch manifest depending on the architecture of the cluster's node running the internal registry.