adm release new with --mirror --from-image-stream should not require a public route for the internal registry
Commit 57ce6b4 introduced the --reference-mode=source option in the oc adm release new command to preserve image location during the creation of a new release image. The author removed a TODO comment regarding support for internal and referential images, although the commit was focusing primarily on referential mode.

Today, if a user sets '--reference-mode' to a value other than 'source', we require the release ImageStream to have an external-facing repository based on the '.Status.PublicDockerImageRepository' field not being empty. The command fails with:

"only image streams or releases with public image repositories can be the source for releases when using the default --reference-mode"

However, users might run this command in a pod, capable of reaching the internal registry url without the need to expose it publicly.
This change introduces logic to check for the presence of the '--mirror' option. If '--mirror' is used, and the 'oc adm release new' command is executed within the cluster itself, there is no longer a need to expose an external route for the internal registry. This adjustment simplifies the process when external access to the internal registry is unnecessary.