pkg/cli/admin/upgrade/status: Add update duration
As a cluster admin, I want to understand how the update is going,
relative to my expectations around duration [1].  If it's waiting on
the Kube API server 5 minutes into the update, that's expected.  If
it's waiting on the Kube API server an hour into the update, that's
unexpected.  There is not a lot of information here, but while it is a
small amount of context, it only consumes a few characters.

  $ OC_ENABLE_CMD_UPGRADE_STATUS=true ./oc adm upgrade status
  An update is in progress for 44m40s: Working towards 4.13.17: 683 of 843 done (81% complete), waiting on network

[1]: https://docs.openshift.com/container-platform/4.14/updating/understanding_updates/understanding-openshift-update-duration.html