oc whoami: Ask from selfsubjectreview and fallback to internal oauth
Currently, `oc whoami` assumes that internal oauth is enabled and that's
why, it fails when an error is gotten from server. However, cluster might
be configured to rely on external OIDC issuer and internal oauth server
might be disabled. Even in that case, `oc whoami` should return the
correct data.

This commit always asks to selfsubjectreview endpoint for whoami first and
it falls back to internal oauth server's user object in case of errors.