Handle invalid related objects with errors
Previously, when  a custom resource had a `status.relatedObject` list
with items that are not valid `ObjectReference`s, then the `inspect`
command encounters an error while trying to gather that resource, and
does not write any of the custom resource's data. Now, it writes the
CR's data even when those errors occur.

Refs:
 - https://issues.redhat.com/browse/ACM-7713

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>