[must-gather] Add support of extra arguments for collection script
This patch adds the ability to pass additional arguments
to the must-gather script. This is an upcoming feature
meant to modularize and optimize the mg collection process.

Note: The current implementation does not forward the args
to the deployed pod when explicitly using a custom entrypoint.

This is done to prevent the args overlap when using custom
collection scripts.

Signed-off-by: black-dragon74 <niryadav@redhat.com>