Include el7 version of oc in 4.11 Fixes rolled out for FIPS make it impossible for a binary compiled on RHELX to run on RHELY (in a FIPS compliant way) if the RHELY host is running with FIPS enabled.
This is because the binary will not find a compatible OpenSSL
library and fall back to using internal Go crypto. Only
OpenSSL is FIPS certified.

With the introduction of "FIPS or Die", binaries run in
FIPS mode but which cannot use OpenSSL will exit with an
error. Customers using RHEL7 with FIPS enabled, therefore,
need a RHEL7 based oc.