pkg/cli/admin/upgrade/rollback: New rollback command
The "which release to roll back to?" criteria are matched to
openshift/cluster-version-operator@12fdd0456b
(pkg/payload/precondition/clusterversion/rollback: Allow previous
version within z-stream, 2023-11-16,
openshift/cluster-version-operator#996), where, based on [1], Lala,
Scott, and Subin all agreed that rollbacks to the cluster's previous
release should not be blocked, as long as the previous and current
version had the same major.minor version (i.e. "within a z-stream",
using x.y.z naming for major.minor.patch).  We agreed to ignore
whether the previous history entry is Completed or Partial, expecting
that most users who are interested in rolling back will make that
decision at the first sign of trouble, and not after attempting a few
roll-forward retargets.

I'm currently blocking Progressing!=False from rolling back, to align
with our CI coverage in openshift/release@f4d37e7f87
(ci-operator/config/openshift/release: Use abort-at=100 for
rollback-oldest-supported, 2023-10-31, openshift/release#45115).  We
could lift this restriction if we wanted to allow rollbacks to be
requested from the middle of an in-progress update, but if we do, we'd
probably want to add matching CI to cover that use-case.

The new subcommand is hidden, as Lala requested in [2]'s Description
on 2021-09-22.  I have not put it behind a feature-gate environment
variable in this commit, but that's also an option if we wanted to
make it even less reachable.

[1]: https://issues.redhat.com/browse/OTA-553
[2]: https://issues.redhat.com/browse/OTA-492