OTA-1087: Add update health section to `adm update status`
Add a section with upgrade health insights to `adm update status`
output, and as a proof-of-concept, populate it with insights about
degraded and unavailable cluster operators during the upgrade.

The CO insights do not report when a CO is in a problematic state for a
short amount of time (5m as an initial value), then a warning insight is
show for them, and once the condition persists over a higher threshold
(20m for unavailable and 40m for degraded), the insight is bumped to
error.

The names, structure and implementation generally follow the concepts
and ideas from the "OpenShift Update Concepts" slides deck, and the
update insights data structure follows [Justin's design](https://github.com/jupierce/oc-update-design-ideas/blob/main/status.md#get-hooksconfigopenshiftiov1alphaupdateinsights),
with minor modifications (and also not fully implemented, some parts are
not needed yet).