Consider kubectl version checks only in cli and cli-artifact images
Currently, when we bump oc to newer k8s versions, our CI jobs are
failing due to kubectl version mismatch error. The reason is that when we bump
oc to newer kubectl(i.e. 1.29.0), cli and cli-artifact images
manifest this version. On the other hand, other images in other operators
have not-bumped kubectl version(i.e. 1.28.2).

This PR only populates kubectl versions in cli and cli-artifact
images(because they are the source of truth) but in addition to that
this PR checks the other images as well to have a consistency with
a rule about all images should match either kubectl version in cli,
cli-artifact or match with each other (all versions in other images
at least should equal to 1.28.2).