pkg/cli/admin/inspectalerts/OWNERS: Assign to monitoring team
Simon, internally:

> ...let's go with the CMO owners...

So this brings in the approver set from [1].

That has the same people as reviewers and approvers, so I'm only
creating a single monitoring-approvers here to ease future
membership-management.  If, in the future, monitoring decides to have
distinct sets for reviewers and approvers, it would be straightforward
to add a new alias.

[1]: https://github.com/openshift/cluster-monitoring-operator/blob/b7e3f50875f2bb1fed912b23fb80a101d3a786c0/OWNERS