Print log rather than returning an error when kubectl version mistmaches
There are images that are based on cli image in the release payload and
those images also manifest their kubectl versions (i.e. cluster-samples-operator).
It is reasonable to assume that all the kubectl versions should match,
because base image is the same for all of them. However, rebuilt strategy
of some images simply rely on a new commit in their repository.
That means if there is no new commit merged for a repository that the image
is residing in, this image will not be rebuilt and still manifests an old
kubectl version. That causes a new release payload generation mechanism a failure.

Due to this uncontrolled environment for kubectl versions, this PR switches
the kubectl version check from error to information log. Users will still
see mismatches but this just won't cause any failure.
This shouldn't create a problem because our source of truths are still
cli and cli-artifacts images.