cli/admin/release: add github method for changelogs
This PR adds flags and logic for generating changelogs using the github
API. When using an OAuth token, this method is considerably faster than
using the `git` method without an existing `git` directory and requires
no disk space. It is considerably slower than using a warmed up `git`
directory, but this may be reduced by using a github cache server such
as the one used by prow.  This new changelog generation method is
intended to be used mainly by the `release-controller`, which will have
access to Openshift Prow's github cache.