adm update status: Add MCO logic needed for worker node status
The needed MCO logic for worker node status is copied over from the
machine-config-operator repository [1]. The required files are copied
over and slightly modified if needed. This is done because vendoring the
original packages from [1] results in massive vendoring changes due to
a small experimental functionality that will be eventually removed and
carried over to a server.

The alternative is to vendor the original packages and introduce the
massive vendoring changes.

[1] https://github.com/openshift/machine-config-operator/tree/11d5151a784c7d4be5255ea41acfbf5092eda592