pkg/cli/admin/release/git: fetch with blob:none
Replace all use of `ensureRemoteForRepo` with
`ensureFetchedRemoteForRepo` and add `--filter=blob:none` to its fetch
command. This reduces disk space used by repos with multiple remotes
added. This also changes the `Clone` function to use the
`remoteNameForRepo` as the origin during to clone to prevent having the
origin be added as a remote twice on multiple runs.