cli/admin/release/git: rename origin after clone
This PR removes the `--origin` flag from the initial git clone as it is
incompatible when used with the `--bare` flag in the version of git in
RHEL 8 and thus the current 4.16 cli and tools images. Instead, we now
rename the `origin` remote to the proper hashed name after the initial
clone finishes.