pkg/cli/admin/upgrade/status: Move Failing from free-form to updateInsight
Structure this output, so it gets all the usual pretty-printing,
detail, etc. that the sad-ClusterOperator conditions are getting.
Sometimes Failing will complain about sad ClusterOperators, and in
that case we'll double up on that messaging.  But we're punting on
"consolidate when multiple updateInsights complain about the same root
cause" for now.  And sometimes Failing will complain about other
resources, such as ProgressDeadlineExceeded operator Deployments [1],
and in that case the information is only flowing out through
ClusterVersion, and not via the other resources we check when
rendering status.

The links to ClusterOperatorDegraded are because [2] folded
Failing=True into ClusterOperatorDegraded alerting, although we still
need to update the runbook to address that change.

The *output updates are via:

  $ go build ./cmd/oc
  $ export  OC_ENABLE_CMD_UPGRADE_STATUS=true
  $ for X in pkg/cli/admin/upgrade/status/examples/*-cv.yaml; do ./oc adm upgrade status --mock-clusterversion "${X}" > "${X/-cv.yaml/.output}"; ./oc adm upgrade status --detailed=all --mock-clusterversion "${X}" > "${X/-cv.yaml/.detailed-output}"; done

[1]: https://github.com/openshift/cluster-version-operator/blob/1acac06742fb0e3e49ffe2294864007f26a7799d/lib/resourcebuilder/apps.go#L122C124-L122C148
[2]: https://github.com/openshift/cluster-version-operator/pull/746