OCPBUGS-33898: `adm update status`: Show information about the upgrade target version (#1767)
* adm update status: Show information about upgrade target version

This commit will show additional information regarding the upgrade.

A new line in the control plane section shows information regarding the
upgrade target version and the previous version.

In the event that the previous version was an incomplete update
(a partial update) and the control plane is currently upgrading, a new
insight is generated, as this fact may influence the upgrade.

* adm update status: Fix target version field for install.