pkg/cli/admin/mustgather: don't manually delete cluster role binding
instead, set the cluster role binding owner reference to the temporary
must-gather namespace. when the namespace is deleted, kubernetes handles
the deletion of the cluster role binding for us.

in cases where must-gather runs ends in failure, this change ensures
that once users manually delete the namespace, the cluster role binding
is also deleted.