`upgrade status`: estimate better via last observed progress
The estimate is based on the projection of "X amount of progress was
achieved in Y amount of time". Previously the "now" (when the command
was run) was used as the amount of time taken to achieve the progress,
which makes the projection inherently diverging: the estimate is always
just prologing.

Instead, try detecting when the last progress was observed. Progress is
assumed to be either last outdated CO that went `Progressing=True`
(assuming it started to update after something else finished) or last
updated CO that went `Progressing=False` (assuming it finished
updating). COs do not use `Progressing` condition to only communicate
they are updating, so this is not always precise, but better than
before.