`update status`: Avoid projecting from small values
Very short durations are very imprecise to project from, so we only
start to project once we have a solid last observable progress duration
(I chose 5 minutes). Early in the update the last observable progress is
likely to be imprecise because of operators not using `Progressing`
condition consistently to communicate that they are updating so a
projection from that will amplify the imprecision.